Moved startDate, endDate and projectCode from project to Batch
