+++ src/resources/batches/BatchShow.tsx
@@ -7,7 +7,8 @@ import {
   TopToolbar,
   FilterButton,
   SelectColumnsButton,
-  DeleteButton
+  DeleteButton,
+  DateField
 } from 'react-admin'
 import { useParams } from 'react-router-dom'
 import * as constants from '../../constants'
@@ -115,6 +116,19 @@ export default function BatchShow(): React.ReactElement {
             <StyledFieldWithLabel label='Receipt notes' source='receiptNotes' />
           </FlexBox>
           <FlexBox>
+            <StyledFieldWithLabel
+              component={DateField}
+              label='Start Date'
+              source='startDate'
+            />
+            <StyledFieldWithLabel
+              component={DateField}
+              label='End Date'
+              source='endDate'
+            />
+          </FlexBox>
+          <FlexBox>
+            <StyledFieldWithLabel label='Project Code' source='projectCode' />
             <StyledFieldWithLabel label='Created' source='createdAt' />
           </FlexBox>
         </TabbedShowLayout.Tab>
